fix(js): remove import cycle
